A:1,清空缓存 2,重置交易密码和用户密码 3,完善请求数据下拉刷新

+++ app/src/main/java/com/yundian/star/been/UserinfoBean.java
@@ -15,16 +15,16 @@ public class UserinfoBean implements Parcelable {
      * type : 0
      */
 
-    private int balance;
+    private double balance;
     private int id;
     private String phone;
     private int type;
 
-    public int getBalance() {
+    public double getBalance() {
         return balance;
     }
 
-    public void setBalance(int balance) {
+    public void setBalance(double balance) {
         this.balance = balance;
     }
 
@@ -59,7 +59,7 @@ public class UserinfoBean implements Parcelable {
 
     @Override
     public void writeToParcel(Parcel dest, int flags) {
-        dest.writeInt(this.balance);
+        dest.writeDouble(this.balance);
         dest.writeInt(this.id);
         dest.writeString(this.phone);
         dest.writeInt(this.type);
